flask-psql/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
flask-cors
